Add staging deployment to config
