+++ scripts/config/local.json
@@ -22,6 +22,6 @@
     "gro-usdc-l2-account": "",
     "gro-referral": "0x2F6579F6031E85b8862848b90118Bb12c0ea5a31",
     "aave-l2-account": "0xc947543144e433c41459b09a765572ee0bf9033e",
-    "aave-swapper": "0xD5ac451B0c50B9476107823Af206eD814a2e2580"
+    "aave-swapper": "0xf953b3A269d80e3eB0F2947630Da976B896A8C5b"
   }
 }
