+++ scripts/config/staging.json
@@ -23,6 +23,7 @@
     "gro-usdc-swapper": "0x5093E5080c89A247Bd9296f4638Ed94d78c76341",
     "gro-usdc-l2-account": "0x59337e7eaff88de66c50fd2f76b02ed69eccdb60",
     "gro-referral": "0x0000000000000000000000000000000000000000",
-    "aave-l2-account": "0xc947543144e433c41459b09a765572ee0bf9033e"
+    "aave-l2-account": "0xc947543144e433c41459b09a765572ee0bf9033e",
+    "aave-swapper": "0xf9A842E176261d87826038d9b2B8d01AD94a56BD"
   }
 }
