+++ scripts/send-funds-to-zksync.ts
@@ -1,19 +1,17 @@
-const hre = require("hardhat");
-const { BigNumber } = require("@ethersproject/bignumber");
+import hre, { ethers } from "hardhat";
 const ConfigLoader = require("./utils/configurator-loader.js");
 
-const { ethers } = hre;
 const config = new ConfigLoader(hre.network.name).load();
 
-const maxFeePerGas = ethers.utils.parseUnits("130", "gwei"); // "base fee + priority fee" on blocknative
+const maxFeePerGas = ethers.utils.parseUnits("100", "gwei"); // "base fee + priority fee" on blocknative
 const maxPriorityFeePerGas = ethers.utils.parseUnits("2", "gwei"); // "max fee" on blocknative
+const options = { maxFeePerGas, maxPriorityFeePerGas };
 
 (async () => {
   try {
-    const signer = await ethers.getSigner();
-    console.log(`Signer is ${signer.address}`);
+    const [signer] = await ethers.getSigners();
     const balance = await ethers.provider.getBalance(signer.address);
-    console.log(`Signer ETH balance is: ${ethers.utils.formatEther(balance)}`);
+    console.log(`Signer is ${signer.address} holding ETH ${ethers.utils.formatEther(balance)}`);
 
     const zkSync = await ethers.getContractAt("IZkSync", config.zkSync);
 
@@ -29,7 +27,7 @@ const maxPriorityFeePerGas = ethers.utils.parseUnits("2", "gwei"); // "max fee" 
     const dai = await ethers.getContractAt("IERC20", await yvDai.token());
     const usdc = await ethers.getContractAt("IERC20", await yvUsdc.token());
 
-    const l2Account = config.argent["yearn-l2-account"];
+    const l2Account = config.argent["aave-l2-account"];
 
     const daiAmount = ethers.utils.parseEther("1000");
     // const daiAmount = BigNumber.from(2).pow(252);
@@ -43,20 +41,22 @@ const maxPriorityFeePerGas = ethers.utils.parseUnits("2", "gwei"); // "max fee" 
     console.log(`usdcAmount   ${usdcAmount}`);
 
     let tx, estimation;
-    // tx = await dai.approve(config.zkSync, daiAmount, { maxFeePerGas, maxPriorityFeePerGas });
-    // console.log(`dai approve hash ${tx.hash}`);
-    // estimation = await zkSync.estimateGas.depositERC20(dai.address, daiAmount, l2Account, { maxFeePerGas, maxPriorityFeePerGas });
-    // console.log(`gas estimation ${estimation}`);
-    tx = await zkSync.depositERC20(dai.address, daiAmount, l2Account, { maxFeePerGas, maxPriorityFeePerGas });
-    console.log(`hash ${tx.hash}`);
 
-    // tx = await usdc.approve(config.zkSync, usdcAmount, { maxFeePerGas, maxPriorityFeePerGas });
-    // console.log(`usdc approve hash ${tx.hash}`);
-    // estimation = await zkSync.estimateGas.depositERC20(usdc.address, usdcAmount, l2Account, { maxFeePerGas, maxPriorityFeePerGas });
+    // DAI:
+    // tx = await dai.approve(config.zkSync, daiAmount, options);
+    // console.log(`dai approve hash ${tx.hash}`);
+    // estimation = await zkSync.estimateGas.depositERC20(dai.address, daiAmount, l2Account, options);
     // console.log(`gas estimation ${estimation}`);
-    tx = await zkSync.depositERC20(usdc.address, usdcAmount, l2Account, { maxFeePerGas, maxPriorityFeePerGas });
-    console.log(`hash ${tx.hash}`);
-    return;
+    // tx = await zkSync.depositERC20(dai.address, daiAmount, l2Account, options);
+    // console.log(`hash ${tx.hash}`);
+
+    // USDC:
+    // tx = await usdc.approve(config.zkSync, usdcAmount, options);
+    // console.log(`usdc approve hash ${tx.hash}`);
+    // estimation = await zkSync.estimateGas.depositERC20(usdc.address, usdcAmount, l2Account, options);
+    // console.log(`gas estimation ${estimation}`);
+    // tx = await zkSync.depositERC20(usdc.address, usdcAmount, l2Account, options);
+    // console.log(`hash ${tx.hash}`);
 
 
     // for ETH:
